+++ src/controleAcces/journalPOATie.java
@@ -73,7 +73,7 @@ public class journalPOATie extends journalPOA
     /**
      * Operation loguerInconnu
      */
-    public void loguerInconnu(int empreinteIdl, controleAcces.journalPackage.date dateHeure, int idZone, int idPorte, String statut, int type)
+    public void loguerInconnu(long empreinteIdl, controleAcces.journalPackage.date dateHeure, int idZone, int idPorte, String statut, int type)
     {
         _tie.loguerInconnu( empreinteIdl,  dateHeure,  idZone,  idPorte,  statut,  type);
     }
